+++ CHANGELOG.md
@@ -147,6 +147,7 @@
 - Accounting tool: removed nowrapfilter from PDF report templates.
 - Etherscan parser/merger: made generic so will support any Etherscan clone.
 - Accounting tool: Lost transactions can now include an optional fee.
+- Accounting tool: use different default filenames for each type of PDF report.
 ### Removed
 - Conversion tool: removed merge parser for Coinbase/Coinbase Pro.
 - Conversion tool: removed filename "is a directory" message.